# api/routes/extract.py
# api/routes/extract.py
"""Extract 能力入口（#32 v3.1 / ADR-0009）。

- POST /extract/jobs：一次提交 = 每个文档一个 Extract Job
  已发布配置 pin Revision；草稿配置冻结 Job 级执行快照
- GET /documents/{id}/extract-result：文档最新正式 Extract 结果
- GET /jobs/{id}/extract-result：指定 Job 的正式 Extract 结果（不回退文档最新）

本轮无 Idempotency-Key / 受理台账（随 #18）。
"""

from typing import Any, Dict, List, Optional

from fastapi import APIRouter, Depends, File, Form, HTTPException, Request, UploadFile
from pydantic import BaseModel, ConfigDict, Field

from api.dependencies.auth import CurrentUser, get_current_user, require_platform_scope
from api.exceptions import AuthorizationError
from api.jobs import create_job, get_job
from api.routes.jobs import _can_access_job
from api.routes.documents.query import _check_document_access
from services.configuration_service import configuration_service
from services.extract_service import (
    ExtractFailure,
    build_execution_spec,
    resolve_extract_spec,
)
from services.result_service import result_service
from services.extract_result_view import resolve_result_view
from services.supabase_service import supabase_service
from services.template_contract import is_canonical_schema, project_fields
from api.routes.parse import (
    _delete_uploaded_document,
    reject_forbidden_direct_fields,
    save_uploaded_document,
)

router = APIRouter(tags=["抽取能力"])

EXTRACT_SAMPLE_KEY = "extract"


class CreateExtractRequest(BaseModel):
    model_config = ConfigDict(extra="forbid")

    template_code: str = Field(min_length=1, max_length=50)
    document_ids: List[str] = Field(min_length=1)


async def _load_definition_and_spec(
    configuration: Dict[str, Any],
) -> tuple[Optional[str], Dict[str, Any]]:
    """返回 (revision_id, execution_spec)；两者按配置状态二选一。

    快照原样保留 target/data_schema 的取值与存在性，非法显式值在这里被
    resolve_extract_spec 拒绝（422、零 Job）；不支持 fields 回退。
    """
    if configuration.get("status") == "published":
        revision_id = configuration.get("current_revision_id")
        revision = await configuration_service.get_revision(str(revision_id or ""))
        if not revision:
            raise HTTPException(status_code=409, detail="配置没有可用的已发布修订")
        spec = build_execution_spec(revision.get("definition") or {})
        resolved = resolve_extract_spec({"execution_spec": spec})
        return revision["id"], spec

    spec = build_execution_spec(configuration.get("draft_definition") or {})
    resolved = resolve_extract_spec({"execution_spec": spec})
    return None, spec


@router.post("/extract/jobs", status_code=201, include_in_schema=False)
async def create_extract_jobs(
    request: CreateExtractRequest,
    user: CurrentUser = Depends(get_current_user),
):
    if not user.tenant_id:
        raise AuthorizationError("请求没有有效的租户范围")
    configuration = await configuration_service.get_published_extract_configuration_by_code(
        user.tenant_id, request.template_code.strip()
    )
    if not configuration:
        raise HTTPException(status_code=404, detail="抽取模板不存在")
    if not user.can_access_tenant(configuration["tenant_id"]):
        raise AuthorizationError("无权使用该 Configuration")
    if configuration.get("type") != "extract":
        raise HTTPException(status_code=409, detail="该配置不是 Extract 类型")
    if configuration.get("status") == "archived":
        raise HTTPException(status_code=409, detail="配置已归档，不能创建新任务")

    document_ids = sorted({str(item).strip() for item in request.document_ids if str(item).strip()})
    if not document_ids:
        raise HTTPException(status_code=422, detail="至少需要一个文档")

    for document_id in document_ids:
        document = await supabase_service.get_document(document_id)
        if not document or document.get("tenant_id") != configuration["tenant_id"]:
            raise HTTPException(status_code=404, detail=f"文档不存在: {document_id}")
        if not user.is_tenant_admin() and document.get("user_id") != user.user_id:
            raise HTTPException(status_code=404, detail=f"文档不存在: {document_id}")

    try:
        revision_id, spec = await _load_definition_and_spec(configuration)
    except ExtractFailure as exc:
        status = 422 if exc.reason in ("schema_invalid", "schema_missing", "target_not_supported") else 409
        raise HTTPException(status_code=status, detail=f"{exc.reason}: {exc.message}")

    job_ids: List[str] = []
    for document_id in document_ids:
        job_id = await create_job(
            created_by=user.user_id,
            related_document_ids=[document_id],
            tenant_id=configuration["tenant_id"],
            configuration_revision_id=revision_id,
            execution_spec=None if revision_id else spec,
        )
        job_ids.append(job_id)

    return {
        "success": True,
        "data": {
            "job_ids": job_ids,
            "template_code": request.template_code.strip(),
            "revision_id": revision_id,
            "mode": "published" if revision_id else "draft",
        },
    }


@router.post("/extract", status_code=202)
async def create_direct_extract(
    request: Request,
    file: UploadFile = File(...),
    template_code: str = Form(..., min_length=1, max_length=50),
    user: CurrentUser = Depends(require_platform_scope("extract")),
):
    """Direct platform capability: upload one file and enqueue Extract.

    ``template_code`` is intentionally the only public template selector.  A
    configuration UUID, revision UUID, tenant, or department field is never
    accepted on this endpoint.
    """
    await reject_forbidden_direct_fields(request)
    if not user.tenant_id:
        raise AuthorizationError("请求没有有效的租户范围")
    code = template_code.strip()
    if not code:
        raise HTTPException(status_code=422, detail="template_code 不能为空")

    configuration = await configuration_service.get_published_extract_configuration_by_code(
        user.tenant_id, code
    )
    if not configuration:
        raise HTTPException(status_code=404, detail="抽取模板不存在")

    uploaded = await save_uploaded_document(file, user)
    try:
        revision_id, _spec = await _load_definition_and_spec(configuration)
        if not revision_id:
            # The strict resolver above requires a published revision.  Keep
            # this guard in case a mocked/legacy store returns malformed data.
            raise HTTPException(status_code=409, detail="配置没有可用的已发布修订")
        job_id = await create_job(
            created_by=user.user_id,
            related_document_ids=[uploaded["document_id"]],
            tenant_id=user.tenant_id,
            configuration_revision_id=revision_id,
            execution_spec=None,
        )
    except HTTPException:
        await _delete_uploaded_document(uploaded["document_id"], uploaded["file_path"])
        raise
    except ExtractFailure as exc:
        await _delete_uploaded_document(uploaded["document_id"], uploaded["file_path"])
        status = 422 if exc.reason in ("schema_invalid", "schema_missing", "target_not_supported") else 409
        raise HTTPException(status_code=status, detail=f"{exc.reason}: {exc.message}")
    except Exception:
        await _delete_uploaded_document(uploaded["document_id"], uploaded["file_path"])
        raise

    return {
        "success": True,
        "data": {
            "document_id": uploaded["document_id"],
            "job_id": job_id,
            "job_ids": [job_id],
            "template_code": code,
            "revision_id": revision_id,
        },
    }


async def _result_response(row: Dict[str, Any], job: Optional[Dict[str, Any]], user: CurrentUser):
    """Add a read-only display projection; never hide raw data on projection errors."""
    try:
        if not job or not await _can_access_job(job, user):
            view = {"status": "unavailable", "reason": "job_unavailable"}
        else:
            view = await resolve_result_view(job, row, configuration_service.get_revision)
    except Exception:
        # Schema/metadata retrieval is separate from access to the already-authorized result.
        view = {"status": "unavailable", "reason": "view_load_failed"}
    template_data = None
    if view.get("status") == "available" and is_canonical_schema(view.get("schema") or {}):
        data = row.get("data")
        if isinstance(data, dict):
            template_data = project_fields(data)
        elif isinstance(data, list):
            template_data = [project_fields(item) for item in data if isinstance(item, dict)]
    return {
        "success": True,
        "result_id": row.get("id"),
        "job_id": row.get("job_id"),
        "config_revision_id": row.get("config_revision_id"),
        "data": row.get("data"),
        "template_data": template_data,
        "engine": row.get("engine"),
        "view": view,
    }


@router.get("/documents/{document_id}/extract-result")
async def get_document_extract_result(
    document_id: str,
    user: CurrentUser = Depends(get_current_user),
):
    document = await supabase_service.get_document(document_id)
    if not document:
        raise HTTPException(status_code=404, detail="文档不存在")
    _check_document_access(document, user, document_id)

    rows = await result_service.list_results(
        document_id=document_id,
        tenant_id=document.get("tenant_id"),
        sample_key=EXTRACT_SAMPLE_KEY,
        limit=1,
    )
    if not rows:
        raise HTTPException(status_code=404, detail="抽取结果不存在")
    row = rows[0]
    # The document is already authorized. Failure to load display metadata is not
    # failure to retrieve its raw result.
    try:
        job = await get_job(str(row.get("job_id") or ""))
    except Exception:
        job = None
    return await _result_response(row, job, user)


@router.get("/jobs/{job_id}/extract-result")
async def get_job_extract_result(
    job_id: str,
    user: CurrentUser = Depends(get_current_user),
):
    job = await get_job(job_id)
    if not job or not await _can_access_job(job, user):
        raise HTTPException(status_code=404, detail="任务不存在")

    rows = await result_service.list_results(
        job_id=job_id,
        tenant_id=job.get("tenant_id"),
        sample_key=EXTRACT_SAMPLE_KEY,
        limit=1,
    )
    if not rows:
        raise HTTPException(status_code=404, detail="该任务没有正式抽取结果")
    return await _result_response(rows[0], job, user)