# openapi.json
{
  "components": {
    "schemas": {
      "Body_create_direct_extract_api_extract_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "template_code": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Template Code",
            "type": "string"
          }
        },
        "required": [
          "file",
          "template_code"
        ],
        "title": "Body_create_direct_extract_api_extract_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "NeoFlow 智能文档处理平台",
    "title": "NeoFlow",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/documents/{document_id}/extract-result": {
      "get": {
        "operationId": "get_document_extract_result_api_documents__document_id__extract_result_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Document Extract Result",
        "tags": [
          "抽取能力",
          "抽取能力"
        ]
      }
    },
...
    "/api/extract": {
      "post": {
        "description": "Direct platform capability: upload one file and enqueue Extract.\n\n``template_code`` is intentionally the only public template selector.  A\nconfiguration UUID, revision UUID, tenant, or department field is never\naccepted on this endpoint.",
        "operationId": "create_direct_extract_api_extract_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-ai-center-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ai-Center-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "x-ai-center-application-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ai-Center-Application-Id"
            }
          },
          {
            "in": "header",
            "name": "x-ai-center-application-version-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ai-Center-Application-Version-Id"
            }
          },
          {
            "in": "header",
            "name": "x-ai-center-invocation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ai-Center-Invocation-Id"
            }
          },
          {
            "in": "header",
            "name": "x-ai-center-caller-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ai-Center-Caller-Id"
            }
          },
          {
            "in": "header",
            "name": "x-ai-center-scope",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ai-Center-Scope"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_direct_extract_api_extract_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Direct Extract",
        "tags": [
          "抽取能力",
          "抽取能力"
        ]
      }
    },
...
    "/api/jobs/{job_id}/extract-result": {
      "get": {
        "operationId": "get_job_extract_result_api_jobs__job_id__extract_result_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Extract Result",
        "tags": [
          "抽取能力",
          "抽取能力"
        ]
      }
    },
...
  }
}